split out splitLink test to new file
